refactor: the engine can be bundled for a browser (#252)
Groundwork for mounting the editor without a Node process. No behaviour
change; every one of the 1504 tests passes untouched.

`compiler.ts` and `projection.ts` loaded Ajv through `createRequire`, which a
bundler cannot follow: importing either put `(0, cre.createRequire)(...)` into
the browser bundle, where it is not a function. That is the incident
`visual-app-browser-safety.test.ts` was written for, and it is why the engine
has been unreachable from a browser. The reason for `createRequire` was never a
Node capability - the comment says so - it was a CJS default-shape disagreement
between this repo's two tsconfigs. A static import with the same `.default ??
module` fallback answers that without the Node API.

The other two Ajv sites take the same fallback. A bare `.default` is correct
only under NodeNext; `apply-command.ts` is reachable from a browser now and
would have read `undefined` once bundled.

`apply-command.ts` also imported `node:fs` and `node:path` at module scope for
one function. A top-level `node:` import is a bundle failure whether or not the
function using it is ever called, so `runApplyCommand` moves to `apply-cli.ts`
with them - the same split `nesting.ts` and `kind-label.ts` already make, for
the same reason. `writeFileSync` was imported and never used. `posixDirectoryOf`
splits on both separators by hand rather than asking `node:path` which platform
it is on, because in a browser there is no answer and a manifest path is
`/`-separated wherever it is read.

Three comments claiming the browser cannot import these modules are now wrong
and say so: the splits still earn their place, on bundle weight rather than on
breakage.
