+++ src/kind-label.ts
@@ -1,8 +1,9 @@
 /**
  * A kind identity is `<profile>#<local id>`, and its label is that local id.
  *
- * This lives apart from `graph-projection.ts` because the browser needs it:
- * the projection imports the compiler, and the compiler reaches for Ajv
- * through `node:module` at load, which a bundle cannot honour.
+ * This lives apart from `graph-projection.ts` because the browser needs it,
+ * and the projection imports the compiler, which loads Ajv and two schemas at
+ * module scope. That is importable from a browser now (#252) and was not
+ * before; it is still a lot of bundle for one label.
  */
 export const kindLabelOf = (kind: string): string => kind.slice(kind.lastIndexOf('#') + 1)
