+++ src/nesting.ts
@@ -2,11 +2,13 @@
  * What a view nests, kept in a module that imports nothing.
  *
  * These live apart from `projection.ts` because the browser needs the value,
- * not only the type, and `projection.ts` reaches for `node:module` to load
- * Ajv. A value import of it from `src/visual-app` therefore pulls
- * `createRequire` into the browser bundle, where it is not a function and the
- * whole app fails to mount (ADR 0101 introduced the constant; this is where it
- * belongs). `projection.ts` re-exports both, so nothing else has to know.
+ * not only the type. `projection.ts` used to reach for `node:module` to load
+ * Ajv, so importing it from `src/visual-app` shipped `createRequire` into the
+ * bundle, where it is not a function and the whole app failed to mount. That
+ * import is static now (#252) and the bundle survives it - but it still drags
+ * Ajv and the projection schema in for one constant, so the split earns its
+ * place on weight rather than on breakage (ADR 0101 introduced the constant;
+ * this is where it belongs). `projection.ts` re-exports both.
  */
 export type NestingKind = 'composition' | 'assignment'
